+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Only write entries that are worth mentioning to users.
 - LLM: Fix Google GenAI provider when thinking toggle is not on
 - UI: Improve approval request wordings
 - Tool: Remove `PatchFile` tool
+- Tool: Rename `Bash`/`CMD` tool to `Shell` tool
 
 ## [0.56] - 2025-11-19